requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
azure-storage-blob
azure-identity
azure-storage-queue